move Instruction(s) and Elem to own module (execution no longer depends on parsing)
